refactor(gateway): swagger and refactor
