--- apps/basket/src/main/resources/application-dev.sample.yaml
@@ -1,17 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:mysql://127.0.0.1:3306/basket?useSSL=false
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    username: root
-    password: 123456
-  flyway:
-    baseline-on-migrate: true
-    baseline-version: 20230307001
-  data:
-    redis:
-      password: 123456
-      host: 127.0.0.1
-      port: 6379
-  devtools:
-    restart:
-      log-condition-evaluation-delta: false
