+++ apps/console-gateway/src/main/java/com/zhaojj11/jam/consolegateway/system/entity/vo/request/LoginUserRequestVO.java
@@ -1,0 +1,14 @@
+package com.zhaojj11.jam.consolegateway.system.entity.vo.request;
+
+import lombok.Getter;
+import lombok.ToString;
+
+/**
+ * @author zhaojj11
+ */
+@Getter
+@ToString
+public class LoginUserRequestVO {
+    private String username;
+    private String password;
+}
