+++ apps/console-gateway/src/main/resources/application.yml
