+++ apps/console-gateway/src/main/resources/db/migration/V20230307001__init_database.sql
@@ -1,0 +1,24 @@
+create table if not exists console_gateway.user
+(
+    id           bigint auto_increment comment '主键'
+        primary key,
+    username     varchar(30)                        null comment '用户名',
+    password     varchar(100)                       null comment '密码',
+    salt         varchar(50)                        null comment '盐值，用于加密',
+    gender       tinyint                            null comment '性别 0-女 1-男',
+    id_card      varchar(30)                        null comment '身份证号',
+    email        varchar(50)                        null comment '邮箱',
+    phone        varchar(30)                        null comment '电话',
+    birthday     datetime                           null comment '生日',
+    age          int                                null comment '年龄',
+    status       tinyint  default 1                 null comment '用户状态——1-正常 2-禁用',
+    last_login   datetime                           null comment '上次登录时间',
+    created_by   bigint                             null comment '创建人的ID',
+    created_time datetime default CURRENT_TIMESTAMP null comment '创建时间',
+    updated_by   bigint                             null comment '修改人ID',
+    updated_time datetime default CURRENT_TIMESTAMP null comment '修改时间',
+    is_deleted   tinyint  default 0                 NOT NULL comment '逻辑删除的状态 0-未删除 1-已删除',
+    remark       varchar(200)                       null comment '备注'
+)
+    comment '用户表';
+
