+++ apps/console-gateway/src/test/resources/application-test.sample.yaml
