+++ apps/console-gateway/src/test/resources/application.yml
